@@ -30,17 +30,17 @@ class TreeView extends InheritedWidget {
   final TreeViewController controller;
 
   /// The tap handler for a node. Passes the node key.
-  final Function(String) onNodeTap;
+  final Function(String)? onNodeTap;
 
   /// Custom builder for nodes. Parameters are the build context and tree node.
-  final Widget Function(BuildContext, Node) nodeBuilder;
+  final Widget Function(BuildContext, Node)? nodeBuilder;
 
   /// The double tap handler for a node. Passes the node key.
-  final Function(String) onNodeDoubleTap;
+  final Function(String)? onNodeDoubleTap;
 
   /// The expand/collapse handler for a node. Passes the node key and the
   /// expansion state.
-  final Function(String, bool) onExpansionChanged;
+  final Function(String, bool)? onExpansionChanged;
 
   /// The theme for [TreeView].
   final TreeViewTheme theme;
@@ -52,7 +52,7 @@ class TreeView extends InheritedWidget {
   final bool allowParentSelect;
 
   /// How the [TreeView] should respond to user input.
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   /// Whether the extent of the [TreeView] should be determined by the contents
   /// being viewed.
@@ -77,8 +77,8 @@ class TreeView extends InheritedWidget {
   final bool supportParentDoubleTap;
 
   TreeView({
-    Key key,
-    @required this.controller,
+    Key? key,
+    required this.controller,
     this.onNodeTap,
     this.onNodeDoubleTap,
     this.physics,
@@ -88,7 +88,7 @@ class TreeView extends InheritedWidget {
     this.shrinkWrap: false,
     this.primary: true,
     this.nodeBuilder,
-    TreeViewTheme theme,
+    TreeViewTheme? theme,
   })  : this.theme = theme ?? const TreeViewTheme(),
         super(
           key: key,
@@ -96,11 +96,11 @@ class TreeView extends InheritedWidget {
             controller,
             shrinkWrap: shrinkWrap,
             primary: primary,
-            physics: physics,
+            physics: physics!,
           ),
         );
 
-  static TreeView of(BuildContext context) =>
+  static TreeView? of(BuildContext context) =>
       context.dependOnInheritedWidgetOfExactType(aspect: TreeView);
 
   @override
@@ -116,9 +116,9 @@ class TreeView extends InheritedWidget {
 
 class _TreeViewData extends StatelessWidget {
   final TreeViewController _controller;
-  final bool shrinkWrap;
-  final bool primary;
-  final ScrollPhysics physics;
+  final bool? shrinkWrap;
+  final bool? primary;
+  final ScrollPhysics? physics;
 
   const _TreeViewData(this._controller,
       {this.shrinkWrap, this.primary, this.physics});
@@ -129,7 +129,7 @@ class _TreeViewData extends StatelessWidget {
     return Theme(
       data: _parentTheme.copyWith(hoverColor: Colors.grey.shade100),
       child: ListView(
-        shrinkWrap: shrinkWrap,
+        shrinkWrap: shrinkWrap!,
         primary: primary,
         physics: physics,
         padding: EdgeInsets.zero,
