@@ -44,9 +44,9 @@ class Utilities {
       int g = 0;
       int b = 0;
       for (var part in parts) {
-        r = int.parse(part.group(1));
-        g = int.parse(part.group(2));
-        b = int.parse(part.group(3));
+        r = int.parse(part.group(1)!);
+        g = int.parse(part.group(2)!);
+        b = int.parse(part.group(3)!);
       }
       return Color.fromARGB(255, r, g, b);
     } else if (_rgbaExp.hasMatch(value)) {
@@ -56,10 +56,10 @@ class Utilities {
       int b = 0;
       double a = 1;
       for (var part in parts) {
-        r = int.parse(part.group(1));
-        g = int.parse(part.group(2));
-        b = int.parse(part.group(3));
-        a = double.parse(part.group(4));
+        r = int.parse(part.group(1)!);
+        g = int.parse(part.group(2)!);
+        b = int.parse(part.group(3)!);
+        a = double.parse(part.group(4)!);
       }
       return Color.fromARGB((255 * a).toInt(), r, g, b);
     } else if (_materialDesignColorExp.hasMatch(value)) {
@@ -663,1984 +663,6 @@ class Utilities {
     final Random _random = Random.secure();
     var values = List<int>.generate(length, (i) => _random.nextInt(256));
     return base64Url.encode(values).substring(0, length);
-  }
-
-  static IconData getIcon(String value) {
-    final String lowered = value == null ? '' : value.toLowerCase();
-    switch (lowered) {
-      case 'ac_unit':
-        return Icons.ac_unit;
-      case 'access_alarm':
-        return Icons.access_alarm;
-      case 'access_alarms':
-        return Icons.access_alarms;
-      case 'access_time':
-        return Icons.access_time;
-      case 'accessibility':
-        return Icons.accessibility;
-      case 'accessibility_new':
-        return Icons.accessibility_new;
-      case 'accessible':
-        return Icons.accessible;
-      case 'accessible_forward':
-        return Icons.accessible_forward;
-      case 'account_balance':
-        return Icons.account_balance;
-      case 'account_balance_wallet':
-        return Icons.account_balance_wallet;
-      case 'account_box':
-        return Icons.account_box;
-      case 'account_circle':
-        return Icons.account_circle;
-      case 'adb':
-        return Icons.adb;
-      case 'add':
-        return Icons.add;
-      case 'add_a_photo':
-        return Icons.add_a_photo;
-      case 'add_alarm':
-        return Icons.add_alarm;
-      case 'add_alert':
-        return Icons.add_alert;
-      case 'add_box':
-        return Icons.add_box;
-      case 'add_call':
-        return Icons.add_call;
-      case 'add_circle':
-        return Icons.add_circle;
-      case 'add_circle_outline':
-        return Icons.add_circle_outline;
-      case 'add_comment':
-        return Icons.add_comment;
-      case 'add_location':
-        return Icons.add_location;
-      case 'add_photo_alternate':
-        return Icons.add_photo_alternate;
-      case 'add_shopping_cart':
-        return Icons.add_shopping_cart;
-      case 'add_to_home_screen':
-        return Icons.add_to_home_screen;
-      case 'add_to_photos':
-        return Icons.add_to_photos;
-      case 'add_to_queue':
-        return Icons.add_to_queue;
-      case 'adjust':
-        return Icons.adjust;
-      case 'airline_seat_flat':
-        return Icons.airline_seat_flat;
-      case 'airline_seat_flat_angled':
-        return Icons.airline_seat_flat_angled;
-      case 'airline_seat_individual_suite':
-        return Icons.airline_seat_individual_suite;
-      case 'airline_seat_legroom_extra':
-        return Icons.airline_seat_legroom_extra;
-      case 'airline_seat_legroom_normal':
-        return Icons.airline_seat_legroom_normal;
-      case 'airline_seat_legroom_reduced':
-        return Icons.airline_seat_legroom_reduced;
-      case 'airline_seat_recline_extra':
-        return Icons.airline_seat_recline_extra;
-      case 'airline_seat_recline_normal':
-        return Icons.airline_seat_recline_normal;
-      case 'airplanemode_active':
-        return Icons.airplanemode_active;
-      case 'airplanemode_inactive':
-        return Icons.airplanemode_inactive;
-      case 'airplay':
-        return Icons.airplay;
-      case 'airport_shuttle':
-        return Icons.airport_shuttle;
-      case 'alarm':
-        return Icons.alarm;
-      case 'alarm_add':
-        return Icons.alarm_add;
-      case 'alarm_off':
-        return Icons.alarm_off;
-      case 'alarm_on':
-        return Icons.alarm_on;
-      case 'album':
-        return Icons.album;
-      case 'all_inclusive':
-        return Icons.all_inclusive;
-      case 'all_out':
-        return Icons.all_out;
-      case 'alternate_email':
-        return Icons.alternate_email;
-      case 'android':
-        return Icons.android;
-      case 'announcement':
-        return Icons.announcement;
-      case 'apps':
-        return Icons.apps;
-      case 'archive':
-        return Icons.archive;
-      case 'arrow_back':
-        return Icons.arrow_back;
-      case 'arrow_back_ios':
-        return Icons.arrow_back_ios;
-      case 'arrow_downward':
-        return Icons.arrow_downward;
-      case 'arrow_drop_down':
-        return Icons.arrow_drop_down;
-      case 'arrow_drop_down_circle':
-        return Icons.arrow_drop_down_circle;
-      case 'arrow_drop_up':
-        return Icons.arrow_drop_up;
-      case 'arrow_forward':
-        return Icons.arrow_forward;
-      case 'arrow_forward_ios':
-        return Icons.arrow_forward_ios;
-      case 'arrow_left':
-        return Icons.arrow_left;
-      case 'arrow_right':
-        return Icons.arrow_right;
-      case 'arrow_upward':
-        return Icons.arrow_upward;
-      case 'art_track':
-        return Icons.art_track;
-      case 'aspect_ratio':
-        return Icons.aspect_ratio;
-      case 'assessment':
-        return Icons.assessment;
-      case 'assignment':
-        return Icons.assignment;
-      case 'assignment_ind':
-        return Icons.assignment_ind;
-      case 'assignment_late':
-        return Icons.assignment_late;
-      case 'assignment_return':
-        return Icons.assignment_return;
-      case 'assignment_returned':
-        return Icons.assignment_returned;
-      case 'assignment_turned_in':
-        return Icons.assignment_turned_in;
-      case 'assistant':
-        return Icons.assistant;
-      case 'assistant_photo':
-        return Icons.assistant_photo;
-      case 'atm':
-        return Icons.atm;
-      case 'attach_file':
-        return Icons.attach_file;
-      case 'attach_money':
-        return Icons.attach_money;
-      case 'attachment':
-        return Icons.attachment;
-      case 'audiotrack':
-        return Icons.audiotrack;
-      case 'autorenew':
-        return Icons.autorenew;
-      case 'av_timer':
-        return Icons.av_timer;
-      case 'backspace':
-        return Icons.backspace;
-      case 'backup':
-        return Icons.backup;
-      case 'battery_alert':
-        return Icons.battery_alert;
-      case 'battery_charging_full':
-        return Icons.battery_charging_full;
-      case 'battery_full':
-        return Icons.battery_full;
-      case 'battery_std':
-        return Icons.battery_std;
-      case 'battery_unknown':
-        return Icons.battery_unknown;
-      case 'beach_access':
-        return Icons.beach_access;
-      case 'beenhere':
-        return Icons.beenhere;
-      case 'block':
-        return Icons.block;
-      case 'bluetooth':
-        return Icons.bluetooth;
-      case 'bluetooth_audio':
-        return Icons.bluetooth_audio;
-      case 'bluetooth_connected':
-        return Icons.bluetooth_connected;
-      case 'bluetooth_disabled':
-        return Icons.bluetooth_disabled;
-      case 'bluetooth_searching':
-        return Icons.bluetooth_searching;
-      case 'blur_circular':
-        return Icons.blur_circular;
-      case 'blur_linear':
-        return Icons.blur_linear;
-      case 'blur_off':
-        return Icons.blur_off;
-      case 'blur_on':
-        return Icons.blur_on;
-      case 'book':
-        return Icons.book;
-      case 'bookmark':
-        return Icons.bookmark;
-      case 'bookmark_border':
-        return Icons.bookmark_border;
-      case 'border_all':
-        return Icons.border_all;
-      case 'border_bottom':
-        return Icons.border_bottom;
-      case 'border_clear':
-        return Icons.border_clear;
-      case 'border_color':
-        return Icons.border_color;
-      case 'border_horizontal':
-        return Icons.border_horizontal;
-      case 'border_inner':
-        return Icons.border_inner;
-      case 'border_left':
-        return Icons.border_left;
-      case 'border_outer':
-        return Icons.border_outer;
-      case 'border_right':
-        return Icons.border_right;
-      case 'border_style':
-        return Icons.border_style;
-      case 'border_top':
-        return Icons.border_top;
-      case 'border_vertical':
-        return Icons.border_vertical;
-      case 'branding_watermark':
-        return Icons.branding_watermark;
-      case 'brightness_1':
-        return Icons.brightness_1;
-      case 'brightness_2':
-        return Icons.brightness_2;
-      case 'brightness_3':
-        return Icons.brightness_3;
-      case 'brightness_4':
-        return Icons.brightness_4;
-      case 'brightness_5':
-        return Icons.brightness_5;
-      case 'brightness_6':
-        return Icons.brightness_6;
-      case 'brightness_7':
-        return Icons.brightness_7;
-      case 'brightness_auto':
-        return Icons.brightness_auto;
-      case 'brightness_high':
-        return Icons.brightness_high;
-      case 'brightness_low':
-        return Icons.brightness_low;
-      case 'brightness_medium':
-        return Icons.brightness_medium;
-      case 'broken_image':
-        return Icons.broken_image;
-      case 'brush':
-        return Icons.brush;
-      case 'bubble_chart':
-        return Icons.bubble_chart;
-      case 'bug_report':
-        return Icons.bug_report;
-      case 'build':
-        return Icons.build;
-      case 'burst_mode':
-        return Icons.burst_mode;
-      case 'business':
-        return Icons.business;
-      case 'business_center':
-        return Icons.business_center;
-      case 'cached':
-        return Icons.cached;
-      case 'cake':
-        return Icons.cake;
-      case 'calendar_today':
-        return Icons.calendar_today;
-      case 'calendar_view_day':
-        return Icons.calendar_view_day;
-      case 'call':
-        return Icons.call;
-      case 'call_end':
-        return Icons.call_end;
-      case 'call_made':
-        return Icons.call_made;
-      case 'call_merge':
-        return Icons.call_merge;
-      case 'call_missed':
-        return Icons.call_missed;
-      case 'call_missed_outgoing':
-        return Icons.call_missed_outgoing;
-      case 'call_received':
-        return Icons.call_received;
-      case 'call_split':
-        return Icons.call_split;
-      case 'call_to_action':
-        return Icons.call_to_action;
-      case 'camera':
-        return Icons.camera;
-      case 'camera_alt':
-        return Icons.camera_alt;
-      case 'camera_enhance':
-        return Icons.camera_enhance;
-      case 'camera_front':
-        return Icons.camera_front;
-      case 'camera_rear':
-        return Icons.camera_rear;
-      case 'camera_roll':
-        return Icons.camera_roll;
-      case 'cancel':
-        return Icons.cancel;
-      case 'card_giftcard':
-        return Icons.card_giftcard;
-      case 'card_membership':
-        return Icons.card_membership;
-      case 'card_travel':
-        return Icons.card_travel;
-      case 'casino':
-        return Icons.casino;
-      case 'cast':
-        return Icons.cast;
-      case 'cast_connected':
-        return Icons.cast_connected;
-      case 'category':
-        return Icons.category;
-      case 'center_focus_strong':
-        return Icons.center_focus_strong;
-      case 'center_focus_weak':
-        return Icons.center_focus_weak;
-      case 'change_history':
-        return Icons.change_history;
-      case 'chat':
-        return Icons.chat;
-      case 'chat_bubble':
-        return Icons.chat_bubble;
-      case 'chat_bubble_outline':
-        return Icons.chat_bubble_outline;
-      case 'check':
-        return Icons.check;
-      case 'check_box':
-        return Icons.check_box;
-      case 'check_box_outline_blank':
-        return Icons.check_box_outline_blank;
-      case 'check_circle':
-        return Icons.check_circle;
-      case 'check_circle_outline':
-        return Icons.check_circle_outline;
-      case 'chevron_left':
-        return Icons.chevron_left;
-      case 'chevron_right':
-        return Icons.chevron_right;
-      case 'child_care':
-        return Icons.child_care;
-      case 'child_friendly':
-        return Icons.child_friendly;
-      case 'chrome_reader_mode':
-        return Icons.chrome_reader_mode;
-      case 'class_':
-        return Icons.class_;
-      case 'clear':
-        return Icons.clear;
-      case 'clear_all':
-        return Icons.clear_all;
-      case 'close':
-        return Icons.close;
-      case 'closed_caption':
-        return Icons.closed_caption;
-      case 'cloud':
-        return Icons.cloud;
-      case 'cloud_circle':
-        return Icons.cloud_circle;
-      case 'cloud_done':
-        return Icons.cloud_done;
-      case 'cloud_download':
-        return Icons.cloud_download;
-      case 'cloud_off':
-        return Icons.cloud_off;
-      case 'cloud_queue':
-        return Icons.cloud_queue;
-      case 'cloud_upload':
-        return Icons.cloud_upload;
-      case 'code':
-        return Icons.code;
-      case 'collections':
-        return Icons.collections;
-      case 'collections_bookmark':
-        return Icons.collections_bookmark;
-      case 'color_lens':
-        return Icons.color_lens;
-      case 'colorize':
-        return Icons.colorize;
-      case 'comment':
-        return Icons.comment;
-      case 'compare':
-        return Icons.compare;
-      case 'compare_arrows':
-        return Icons.compare_arrows;
-      case 'computer':
-        return Icons.computer;
-      case 'confirmation_number':
-        return Icons.confirmation_number;
-      case 'contact_mail':
-        return Icons.contact_mail;
-      case 'contact_phone':
-        return Icons.contact_phone;
-      case 'contacts':
-        return Icons.contacts;
-      case 'content_copy':
-        return Icons.content_copy;
-      case 'content_cut':
-        return Icons.content_cut;
-      case 'content_paste':
-        return Icons.content_paste;
-      case 'control_point':
-        return Icons.control_point;
-      case 'control_point_duplicate':
-        return Icons.control_point_duplicate;
-      case 'copyright':
-        return Icons.copyright;
-      case 'create':
-        return Icons.create;
-      case 'create_new_folder':
-        return Icons.create_new_folder;
-      case 'credit_card':
-        return Icons.credit_card;
-      case 'crop':
-        return Icons.crop;
-      case 'crop_3_2':
-        return Icons.crop_3_2;
-      case 'crop_5_4':
-        return Icons.crop_5_4;
-      case 'crop_7_5':
-        return Icons.crop_7_5;
-      case 'crop_16_9':
-        return Icons.crop_16_9;
-      case 'crop_din':
-        return Icons.crop_din;
-      case 'crop_free':
-        return Icons.crop_free;
-      case 'crop_landscape':
-        return Icons.crop_landscape;
-      case 'crop_original':
-        return Icons.crop_original;
-      case 'crop_portrait':
-        return Icons.crop_portrait;
-      case 'crop_rotate':
-        return Icons.crop_rotate;
-      case 'crop_square':
-        return Icons.crop_square;
-      case 'dashboard':
-        return Icons.dashboard;
-      case 'data_usage':
-        return Icons.data_usage;
-      case 'date_range':
-        return Icons.date_range;
-      case 'dehaze':
-        return Icons.dehaze;
-      case 'delete':
-        return Icons.delete;
-      case 'delete_forever':
-        return Icons.delete_forever;
-      case 'delete_outline':
-        return Icons.delete_outline;
-      case 'delete_sweep':
-        return Icons.delete_sweep;
-      case 'departure_board':
-        return Icons.departure_board;
-      case 'description':
-        return Icons.description;
-      case 'desktop_mac':
-        return Icons.desktop_mac;
-      case 'desktop_windows':
-        return Icons.desktop_windows;
-      case 'details':
-        return Icons.details;
-      case 'developer_board':
-        return Icons.developer_board;
-      case 'developer_mode':
-        return Icons.developer_mode;
-      case 'device_hub':
-        return Icons.device_hub;
-      case 'device_unknown':
-        return Icons.device_unknown;
-      case 'devices':
-        return Icons.devices;
-      case 'devices_other':
-        return Icons.devices_other;
-      case 'dialer_sip':
-        return Icons.dialer_sip;
-      case 'dialpad':
-        return Icons.dialpad;
-      case 'directions':
-        return Icons.directions;
-      case 'directions_bike':
-        return Icons.directions_bike;
-      case 'directions_boat':
-        return Icons.directions_boat;
-      case 'directions_bus':
-        return Icons.directions_bus;
-      case 'directions_car':
-        return Icons.directions_car;
-      case 'directions_railway':
-        return Icons.directions_railway;
-      case 'directions_run':
-        return Icons.directions_run;
-      case 'directions_subway':
-        return Icons.directions_subway;
-      case 'directions_transit':
-        return Icons.directions_transit;
-      case 'directions_walk':
-        return Icons.directions_walk;
-      case 'disc_full':
-        return Icons.disc_full;
-      case 'dns':
-        return Icons.dns;
-      case 'do_not_disturb':
-        return Icons.do_not_disturb;
-      case 'do_not_disturb_alt':
-        return Icons.do_not_disturb_alt;
-      case 'do_not_disturb_off':
-        return Icons.do_not_disturb_off;
-      case 'do_not_disturb_on':
-        return Icons.do_not_disturb_on;
-      case 'dock':
-        return Icons.dock;
-      case 'domain':
-        return Icons.domain;
-      case 'done':
-        return Icons.done;
-      case 'done_all':
-        return Icons.done_all;
-      case 'done_outline':
-        return Icons.done_outline;
-      case 'donut_large':
-        return Icons.donut_large;
-      case 'donut_small':
-        return Icons.donut_small;
-      case 'drafts':
-        return Icons.drafts;
-      case 'drag_handle':
-        return Icons.drag_handle;
-      case 'drive_eta':
-        return Icons.drive_eta;
-      case 'dvr':
-        return Icons.dvr;
-      case 'edit':
-        return Icons.edit;
-      case 'edit_attributes':
-        return Icons.edit_attributes;
-      case 'edit_location':
-        return Icons.edit_location;
-      case 'eject':
-        return Icons.eject;
-      case 'email':
-        return Icons.email;
-      case 'enhanced_encryption':
-        return Icons.enhanced_encryption;
-      case 'equalizer':
-        return Icons.equalizer;
-      case 'error':
-        return Icons.error;
-      case 'error_outline':
-        return Icons.error_outline;
-      case 'euro_symbol':
-        return Icons.euro_symbol;
-      case 'ev_station':
-        return Icons.ev_station;
-      case 'event':
-        return Icons.event;
-      case 'event_available':
-        return Icons.event_available;
-      case 'event_busy':
-        return Icons.event_busy;
-      case 'event_note':
-        return Icons.event_note;
-      case 'event_seat':
-        return Icons.event_seat;
-      case 'exit_to_app':
-        return Icons.exit_to_app;
-      case 'expand_less':
-        return Icons.expand_less;
-      case 'expand_more':
-        return Icons.expand_more;
-      case 'explicit':
-        return Icons.explicit;
-      case 'explore':
-        return Icons.explore;
-      case 'exposure':
-        return Icons.exposure;
-      case 'exposure_neg_1':
-        return Icons.exposure_neg_1;
-      case 'exposure_neg_2':
-        return Icons.exposure_neg_2;
-      case 'exposure_plus_1':
-        return Icons.exposure_plus_1;
-      case 'exposure_plus_2':
-        return Icons.exposure_plus_2;
-      case 'exposure_zero':
-        return Icons.exposure_zero;
-      case 'extension':
-        return Icons.extension;
-      case 'face':
-        return Icons.face;
-      case 'fast_forward':
-        return Icons.fast_forward;
-      case 'fast_rewind':
-        return Icons.fast_rewind;
-      case 'fastfood':
-        return Icons.fastfood;
-      case 'favorite':
-        return Icons.favorite;
-      case 'favorite_border':
-        return Icons.favorite_border;
-      case 'featured_play_list':
-        return Icons.featured_play_list;
-      case 'featured_video':
-        return Icons.featured_video;
-      case 'feedback':
-        return Icons.feedback;
-      case 'fiber_dvr':
-        return Icons.fiber_dvr;
-      case 'fiber_manual_record':
-        return Icons.fiber_manual_record;
-      case 'fiber_new':
-        return Icons.fiber_new;
-      case 'fiber_pin':
-        return Icons.fiber_pin;
-      case 'fiber_smart_record':
-        return Icons.fiber_smart_record;
-      case 'file_download':
-        return Icons.file_download;
-      case 'file_upload':
-        return Icons.file_upload;
-      case 'filter':
-        return Icons.filter;
-      case 'filter_1':
-        return Icons.filter_1;
-      case 'filter_2':
-        return Icons.filter_2;
-      case 'filter_3':
-        return Icons.filter_3;
-      case 'filter_4':
-        return Icons.filter_4;
-      case 'filter_5':
-        return Icons.filter_5;
-      case 'filter_6':
-        return Icons.filter_6;
-      case 'filter_7':
-        return Icons.filter_7;
-      case 'filter_8':
-        return Icons.filter_8;
-      case 'filter_9':
-        return Icons.filter_9;
-      case 'filter_9_plus':
-        return Icons.filter_9_plus;
-      case 'filter_b_and_w':
-        return Icons.filter_b_and_w;
-      case 'filter_center_focus':
-        return Icons.filter_center_focus;
-      case 'filter_drama':
-        return Icons.filter_drama;
-      case 'filter_frames':
-        return Icons.filter_frames;
-      case 'filter_hdr':
-        return Icons.filter_hdr;
-      case 'filter_list':
-        return Icons.filter_list;
-      case 'filter_none':
-        return Icons.filter_none;
-      case 'filter_tilt_shift':
-        return Icons.filter_tilt_shift;
-      case 'filter_vintage':
-        return Icons.filter_vintage;
-      case 'find_in_page':
-        return Icons.find_in_page;
-      case 'find_replace':
-        return Icons.find_replace;
-      case 'fingerprint':
-        return Icons.fingerprint;
-      case 'first_page':
-        return Icons.first_page;
-      case 'fitness_center':
-        return Icons.fitness_center;
-      case 'flag':
-        return Icons.flag;
-      case 'flare':
-        return Icons.flare;
-      case 'flash_auto':
-        return Icons.flash_auto;
-      case 'flash_off':
-        return Icons.flash_off;
-      case 'flash_on':
-        return Icons.flash_on;
-      case 'flight':
-        return Icons.flight;
-      case 'flight_land':
-        return Icons.flight_land;
-      case 'flight_takeoff':
-        return Icons.flight_takeoff;
-      case 'flip':
-        return Icons.flip;
-      case 'flip_to_back':
-        return Icons.flip_to_back;
-      case 'flip_to_front':
-        return Icons.flip_to_front;
-      case 'folder':
-        return Icons.folder;
-      case 'folder_open':
-        return Icons.folder_open;
-      case 'folder_shared':
-        return Icons.folder_shared;
-      case 'folder_special':
-        return Icons.folder_special;
-      case 'font_download':
-        return Icons.font_download;
-      case 'format_align_center':
-        return Icons.format_align_center;
-      case 'format_align_justify':
-        return Icons.format_align_justify;
-      case 'format_align_left':
-        return Icons.format_align_left;
-      case 'format_align_right':
-        return Icons.format_align_right;
-      case 'format_bold':
-        return Icons.format_bold;
-      case 'format_clear':
-        return Icons.format_clear;
-      case 'format_color_fill':
-        return Icons.format_color_fill;
-      case 'format_color_reset':
-        return Icons.format_color_reset;
-      case 'format_color_text':
-        return Icons.format_color_text;
-      case 'format_indent_decrease':
-        return Icons.format_indent_decrease;
-      case 'format_indent_increase':
-        return Icons.format_indent_increase;
-      case 'format_italic':
-        return Icons.format_italic;
-      case 'format_line_spacing':
-        return Icons.format_line_spacing;
-      case 'format_list_bulleted':
-        return Icons.format_list_bulleted;
-      case 'format_list_numbered':
-        return Icons.format_list_numbered;
-      case 'format_list_numbered_rtl':
-        return Icons.format_list_numbered_rtl;
-      case 'format_paint':
-        return Icons.format_paint;
-      case 'format_quote':
-        return Icons.format_quote;
-      case 'format_shapes':
-        return Icons.format_shapes;
-      case 'format_size':
-        return Icons.format_size;
-      case 'format_strikethrough':
-        return Icons.format_strikethrough;
-      case 'format_textdirection_l_to_r':
-        return Icons.format_textdirection_l_to_r;
-      case 'format_textdirection_r_to_l':
-        return Icons.format_textdirection_r_to_l;
-      case 'format_underlined':
-        return Icons.format_underlined;
-      case 'forum':
-        return Icons.forum;
-      case 'forward':
-        return Icons.forward;
-      case 'forward_5':
-        return Icons.forward_5;
-      case 'forward_10':
-        return Icons.forward_10;
-      case 'forward_30':
-        return Icons.forward_30;
-      case 'four_k':
-        return Icons.four_k;
-      case 'free_breakfast':
-        return Icons.free_breakfast;
-      case 'fullscreen':
-        return Icons.fullscreen;
-      case 'fullscreen_exit':
-        return Icons.fullscreen_exit;
-      case 'functions':
-        return Icons.functions;
-      case 'g_translate':
-        return Icons.g_translate;
-      case 'gamepad':
-        return Icons.gamepad;
-      case 'games':
-        return Icons.games;
-      case 'gavel':
-        return Icons.gavel;
-      case 'gesture':
-        return Icons.gesture;
-      case 'get_app':
-        return Icons.get_app;
-      case 'gif':
-        return Icons.gif;
-      case 'golf_course':
-        return Icons.golf_course;
-      case 'gps_fixed':
-        return Icons.gps_fixed;
-      case 'gps_not_fixed':
-        return Icons.gps_not_fixed;
-      case 'gps_off':
-        return Icons.gps_off;
-      case 'grade':
-        return Icons.grade;
-      case 'gradient':
-        return Icons.gradient;
-      case 'grain':
-        return Icons.grain;
-      case 'graphic_eq':
-        return Icons.graphic_eq;
-      case 'grid_off':
-        return Icons.grid_off;
-      case 'grid_on':
-        return Icons.grid_on;
-      case 'group':
-        return Icons.group;
-      case 'group_add':
-        return Icons.group_add;
-      case 'group_work':
-        return Icons.group_work;
-      case 'hd':
-        return Icons.hd;
-      case 'hdr_off':
-        return Icons.hdr_off;
-      case 'hdr_on':
-        return Icons.hdr_on;
-      case 'hdr_strong':
-        return Icons.hdr_strong;
-      case 'hdr_weak':
-        return Icons.hdr_weak;
-      case 'headset':
-        return Icons.headset;
-      case 'headset_mic':
-        return Icons.headset_mic;
-      case 'headset_off':
-        return Icons.headset_off;
-      case 'healing':
-        return Icons.healing;
-      case 'hearing':
-        return Icons.hearing;
-      case 'help':
-        return Icons.help;
-      case 'help_outline':
-        return Icons.help_outline;
-      case 'high_quality':
-        return Icons.high_quality;
-      case 'highlight':
-        return Icons.highlight;
-      case 'highlight_off':
-        return Icons.highlight_off;
-      case 'history':
-        return Icons.history;
-      case 'home':
-        return Icons.home;
-      case 'hot_tub':
-        return Icons.hot_tub;
-      case 'hotel':
-        return Icons.hotel;
-      case 'hourglass_empty':
-        return Icons.hourglass_empty;
-      case 'hourglass_full':
-        return Icons.hourglass_full;
-      case 'http':
-        return Icons.http;
-      case 'https':
-        return Icons.https;
-      case 'image':
-        return Icons.image;
-      case 'image_aspect_ratio':
-        return Icons.image_aspect_ratio;
-      case 'import_contacts':
-        return Icons.import_contacts;
-      case 'import_export':
-        return Icons.import_export;
-      case 'important_devices':
-        return Icons.important_devices;
-      case 'inbox':
-        return Icons.inbox;
-      case 'indeterminate_check_box':
-        return Icons.indeterminate_check_box;
-      case 'info':
-        return Icons.info;
-      case 'info_outline':
-        return Icons.info_outline;
-      case 'input':
-        return Icons.input;
-      case 'insert_chart':
-        return Icons.insert_chart;
-      case 'insert_comment':
-        return Icons.insert_comment;
-      case 'insert_drive_file':
-        return Icons.insert_drive_file;
-      case 'insert_emoticon':
-        return Icons.insert_emoticon;
-      case 'insert_invitation':
-        return Icons.insert_invitation;
-      case 'insert_link':
-        return Icons.insert_link;
-      case 'insert_photo':
-        return Icons.insert_photo;
-      case 'invert_colors':
-        return Icons.invert_colors;
-      case 'invert_colors_off':
-        return Icons.invert_colors_off;
-      case 'iso':
-        return Icons.iso;
-      case 'keyboard':
-        return Icons.keyboard;
-      case 'keyboard_arrow_down':
-        return Icons.keyboard_arrow_down;
-      case 'keyboard_arrow_left':
-        return Icons.keyboard_arrow_left;
-      case 'keyboard_arrow_right':
-        return Icons.keyboard_arrow_right;
-      case 'keyboard_arrow_up':
-        return Icons.keyboard_arrow_up;
-      case 'keyboard_backspace':
-        return Icons.keyboard_backspace;
-      case 'keyboard_capslock':
-        return Icons.keyboard_capslock;
-      case 'keyboard_hide':
-        return Icons.keyboard_hide;
-      case 'keyboard_return':
-        return Icons.keyboard_return;
-      case 'keyboard_tab':
-        return Icons.keyboard_tab;
-      case 'keyboard_voice':
-        return Icons.keyboard_voice;
-      case 'kitchen':
-        return Icons.kitchen;
-      case 'label':
-        return Icons.label;
-      case 'label_important':
-        return Icons.label_important;
-      case 'label_outline':
-        return Icons.label_outline;
-      case 'landscape':
-        return Icons.landscape;
-      case 'language':
-        return Icons.language;
-      case 'laptop':
-        return Icons.laptop;
-      case 'laptop_chromebook':
-        return Icons.laptop_chromebook;
-      case 'laptop_mac':
-        return Icons.laptop_mac;
-      case 'laptop_windows':
-        return Icons.laptop_windows;
-      case 'last_page':
-        return Icons.last_page;
-      case 'launch':
-        return Icons.launch;
-      case 'layers':
-        return Icons.layers;
-      case 'layers_clear':
-        return Icons.layers_clear;
-      case 'leak_add':
-        return Icons.leak_add;
-      case 'leak_remove':
-        return Icons.leak_remove;
-      case 'lens':
-        return Icons.lens;
-      case 'library_add':
-        return Icons.library_add;
-      case 'library_books':
-        return Icons.library_books;
-      case 'library_music':
-        return Icons.library_music;
-      case 'lightbulb_outline':
-        return Icons.lightbulb_outline;
-      case 'line_style':
-        return Icons.line_style;
-      case 'line_weight':
-        return Icons.line_weight;
-      case 'linear_scale':
-        return Icons.linear_scale;
-      case 'link':
-        return Icons.link;
-      case 'link_off':
-        return Icons.link_off;
-      case 'linked_camera':
-        return Icons.linked_camera;
-      case 'list':
-        return Icons.list;
-      case 'live_help':
-        return Icons.live_help;
-      case 'live_tv':
-        return Icons.live_tv;
-      case 'local_activity':
-        return Icons.local_activity;
-      case 'local_airport':
-        return Icons.local_airport;
-      case 'local_atm':
-        return Icons.local_atm;
-      case 'local_bar':
-        return Icons.local_bar;
-      case 'local_cafe':
-        return Icons.local_cafe;
-      case 'local_car_wash':
-        return Icons.local_car_wash;
-      case 'local_convenience_store':
-        return Icons.local_convenience_store;
-      case 'local_dining':
-        return Icons.local_dining;
-      case 'local_drink':
-        return Icons.local_drink;
-      case 'local_florist':
-        return Icons.local_florist;
-      case 'local_gas_station':
-        return Icons.local_gas_station;
-      case 'local_grocery_store':
-        return Icons.local_grocery_store;
-      case 'local_hospital':
-        return Icons.local_hospital;
-      case 'local_hotel':
-        return Icons.local_hotel;
-      case 'local_laundry_service':
-        return Icons.local_laundry_service;
-      case 'local_library':
-        return Icons.local_library;
-      case 'local_mall':
-        return Icons.local_mall;
-      case 'local_movies':
-        return Icons.local_movies;
-      case 'local_offer':
-        return Icons.local_offer;
-      case 'local_parking':
-        return Icons.local_parking;
-      case 'local_pharmacy':
-        return Icons.local_pharmacy;
-      case 'local_phone':
-        return Icons.local_phone;
-      case 'local_pizza':
-        return Icons.local_pizza;
-      case 'local_play':
-        return Icons.local_play;
-      case 'local_post_office':
-        return Icons.local_post_office;
-      case 'local_printshop':
-        return Icons.local_printshop;
-      case 'local_see':
-        return Icons.local_see;
-      case 'local_shipping':
-        return Icons.local_shipping;
-      case 'local_taxi':
-        return Icons.local_taxi;
-      case 'location_city':
-        return Icons.location_city;
-      case 'location_disabled':
-        return Icons.location_disabled;
-      case 'location_off':
-        return Icons.location_off;
-      case 'location_on':
-        return Icons.location_on;
-      case 'location_searching':
-        return Icons.location_searching;
-      case 'lock':
-        return Icons.lock;
-      case 'lock_open':
-        return Icons.lock_open;
-      case 'lock_outline':
-        return Icons.lock_outline;
-      case 'looks':
-        return Icons.looks;
-      case 'looks_3':
-        return Icons.looks_3;
-      case 'looks_4':
-        return Icons.looks_4;
-      case 'looks_5':
-        return Icons.looks_5;
-      case 'looks_6':
-        return Icons.looks_6;
-      case 'looks_one':
-        return Icons.looks_one;
-      case 'looks_two':
-        return Icons.looks_two;
-      case 'loop':
-        return Icons.loop;
-      case 'loupe':
-        return Icons.loupe;
-      case 'low_priority':
-        return Icons.low_priority;
-      case 'loyalty':
-        return Icons.loyalty;
-      case 'mail':
-        return Icons.mail;
-      case 'mail_outline':
-        return Icons.mail_outline;
-      case 'map':
-        return Icons.map;
-      case 'markunread':
-        return Icons.markunread;
-      case 'markunread_mailbox':
-        return Icons.markunread_mailbox;
-      case 'maximize':
-        return Icons.maximize;
-      case 'memory':
-        return Icons.memory;
-      case 'menu':
-        return Icons.menu;
-      case 'merge_type':
-        return Icons.merge_type;
-      case 'message':
-        return Icons.message;
-      case 'mic':
-        return Icons.mic;
-      case 'mic_none':
-        return Icons.mic_none;
-      case 'mic_off':
-        return Icons.mic_off;
-      case 'minimize':
-        return Icons.minimize;
-      case 'missed_video_call':
-        return Icons.missed_video_call;
-      case 'mms':
-        return Icons.mms;
-      case 'mobile_screen_share':
-        return Icons.mobile_screen_share;
-      case 'mode_comment':
-        return Icons.mode_comment;
-      case 'mode_edit':
-        return Icons.mode_edit;
-      case 'monetization_on':
-        return Icons.monetization_on;
-      case 'money_off':
-        return Icons.money_off;
-      case 'monochrome_photos':
-        return Icons.monochrome_photos;
-      case 'mood':
-        return Icons.mood;
-      case 'mood_bad':
-        return Icons.mood_bad;
-      case 'more':
-        return Icons.more;
-      case 'more_horiz':
-        return Icons.more_horiz;
-      case 'more_vert':
-        return Icons.more_vert;
-      case 'motorcycle':
-        return Icons.motorcycle;
-      case 'mouse':
-        return Icons.mouse;
-      case 'move_to_inbox':
-        return Icons.move_to_inbox;
-      case 'movie':
-        return Icons.movie;
-      case 'movie_creation':
-        return Icons.movie_creation;
-      case 'movie_filter':
-        return Icons.movie_filter;
-      case 'multiline_chart':
-        return Icons.multiline_chart;
-      case 'music_note':
-        return Icons.music_note;
-      case 'music_video':
-        return Icons.music_video;
-      case 'my_location':
-        return Icons.my_location;
-      case 'nature':
-        return Icons.nature;
-      case 'nature_people':
-        return Icons.nature_people;
-      case 'navigate_before':
-        return Icons.navigate_before;
-      case 'navigate_next':
-        return Icons.navigate_next;
-      case 'navigation':
-        return Icons.navigation;
-      case 'near_me':
-        return Icons.near_me;
-      case 'network_cell':
-        return Icons.network_cell;
-      case 'network_check':
-        return Icons.network_check;
-      case 'network_locked':
-        return Icons.network_locked;
-      case 'network_wifi':
-        return Icons.network_wifi;
-      case 'new_releases':
-        return Icons.new_releases;
-      case 'next_week':
-        return Icons.next_week;
-      case 'nfc':
-        return Icons.nfc;
-      case 'no_encryption':
-        return Icons.no_encryption;
-      case 'no_sim':
-        return Icons.no_sim;
-      case 'not_interested':
-        return Icons.not_interested;
-      case 'not_listed_location':
-        return Icons.not_listed_location;
-      case 'note':
-        return Icons.note;
-      case 'note_add':
-        return Icons.note_add;
-      case 'notification_important':
-        return Icons.notification_important;
-      case 'notifications':
-        return Icons.notifications;
-      case 'notifications_active':
-        return Icons.notifications_active;
-      case 'notifications_none':
-        return Icons.notifications_none;
-      case 'notifications_off':
-        return Icons.notifications_off;
-      case 'notifications_paused':
-        return Icons.notifications_paused;
-      case 'offline_bolt':
-        return Icons.offline_bolt;
-      case 'offline_pin':
-        return Icons.offline_pin;
-      case 'ondemand_video':
-        return Icons.ondemand_video;
-      case 'opacity':
-        return Icons.opacity;
-      case 'open_in_browser':
-        return Icons.open_in_browser;
-      case 'open_in_new':
-        return Icons.open_in_new;
-      case 'open_with':
-        return Icons.open_with;
-      case 'outlined_flag':
-        return Icons.outlined_flag;
-      case 'pages':
-        return Icons.pages;
-      case 'pageview':
-        return Icons.pageview;
-      case 'palette':
-        return Icons.palette;
-      case 'pan_tool':
-        return Icons.pan_tool;
-      case 'panorama':
-        return Icons.panorama;
-      case 'panorama_fish_eye':
-        return Icons.panorama_fish_eye;
-      case 'panorama_horizontal':
-        return Icons.panorama_horizontal;
-      case 'panorama_vertical':
-        return Icons.panorama_vertical;
-      case 'panorama_wide_angle':
-        return Icons.panorama_wide_angle;
-      case 'party_mode':
-        return Icons.party_mode;
-      case 'pause':
-        return Icons.pause;
-      case 'pause_circle_filled':
-        return Icons.pause_circle_filled;
-      case 'pause_circle_outline':
-        return Icons.pause_circle_outline;
-      case 'payment':
-        return Icons.payment;
-      case 'people':
-        return Icons.people;
-      case 'people_outline':
-        return Icons.people_outline;
-      case 'perm_camera_mic':
-        return Icons.perm_camera_mic;
-      case 'perm_contact_calendar':
-        return Icons.perm_contact_calendar;
-      case 'perm_data_setting':
-        return Icons.perm_data_setting;
-      case 'perm_device_information':
-        return Icons.perm_device_information;
-      case 'perm_identity':
-        return Icons.perm_identity;
-      case 'perm_media':
-        return Icons.perm_media;
-      case 'perm_phone_msg':
-        return Icons.perm_phone_msg;
-      case 'perm_scan_wifi':
-        return Icons.perm_scan_wifi;
-      case 'person':
-        return Icons.person;
-      case 'person_add':
-        return Icons.person_add;
-      case 'person_outline':
-        return Icons.person_outline;
-      case 'person_pin':
-        return Icons.person_pin;
-      case 'person_pin_circle':
-        return Icons.person_pin_circle;
-      case 'personal_video':
-        return Icons.personal_video;
-      case 'pets':
-        return Icons.pets;
-      case 'phone':
-        return Icons.phone;
-      case 'phone_android':
-        return Icons.phone_android;
-      case 'phone_bluetooth_speaker':
-        return Icons.phone_bluetooth_speaker;
-      case 'phone_forwarded':
-        return Icons.phone_forwarded;
-      case 'phone_in_talk':
-        return Icons.phone_in_talk;
-      case 'phone_iphone':
-        return Icons.phone_iphone;
-      case 'phone_locked':
-        return Icons.phone_locked;
-      case 'phone_missed':
-        return Icons.phone_missed;
-      case 'phone_paused':
-        return Icons.phone_paused;
-      case 'phonelink':
-        return Icons.phonelink;
-      case 'phonelink_erase':
-        return Icons.phonelink_erase;
-      case 'phonelink_lock':
-        return Icons.phonelink_lock;
-      case 'phonelink_off':
-        return Icons.phonelink_off;
-      case 'phonelink_ring':
-        return Icons.phonelink_ring;
-      case 'phonelink_setup':
-        return Icons.phonelink_setup;
-      case 'photo':
-        return Icons.photo;
-      case 'photo_album':
-        return Icons.photo_album;
-      case 'photo_camera':
-        return Icons.photo_camera;
-      case 'photo_filter':
-        return Icons.photo_filter;
-      case 'photo_library':
-        return Icons.photo_library;
-      case 'photo_size_select_actual':
-        return Icons.photo_size_select_actual;
-      case 'photo_size_select_large':
-        return Icons.photo_size_select_large;
-      case 'photo_size_select_small':
-        return Icons.photo_size_select_small;
-      case 'picture_as_pdf':
-        return Icons.picture_as_pdf;
-      case 'picture_in_picture':
-        return Icons.picture_in_picture;
-      case 'picture_in_picture_alt':
-        return Icons.picture_in_picture_alt;
-      case 'pie_chart':
-        return Icons.pie_chart;
-      case 'pie_chart_outlined':
-        return Icons.pie_chart_outlined;
-      case 'pin_drop':
-        return Icons.pin_drop;
-      case 'place':
-        return Icons.place;
-      case 'play_arrow':
-        return Icons.play_arrow;
-      case 'play_circle_filled':
-        return Icons.play_circle_filled;
-      case 'play_circle_outline':
-        return Icons.play_circle_outline;
-      case 'play_for_work':
-        return Icons.play_for_work;
-      case 'playlist_add':
-        return Icons.playlist_add;
-      case 'playlist_add_check':
-        return Icons.playlist_add_check;
-      case 'playlist_play':
-        return Icons.playlist_play;
-      case 'plus_one':
-        return Icons.plus_one;
-      case 'poll':
-        return Icons.poll;
-      case 'polymer':
-        return Icons.polymer;
-      case 'pool':
-        return Icons.pool;
-      case 'portable_wifi_off':
-        return Icons.portable_wifi_off;
-      case 'portrait':
-        return Icons.portrait;
-      case 'power':
-        return Icons.power;
-      case 'power_input':
-        return Icons.power_input;
-      case 'power_settings_new':
-        return Icons.power_settings_new;
-      case 'pregnant_woman':
-        return Icons.pregnant_woman;
-      case 'present_to_all':
-        return Icons.present_to_all;
-      case 'print':
-        return Icons.print;
-      case 'priority_high':
-        return Icons.priority_high;
-      case 'public':
-        return Icons.public;
-      case 'publish':
-        return Icons.publish;
-      case 'query_builder':
-        return Icons.query_builder;
-      case 'question_answer':
-        return Icons.question_answer;
-      case 'queue':
-        return Icons.queue;
-      case 'queue_music':
-        return Icons.queue_music;
-      case 'queue_play_next':
-        return Icons.queue_play_next;
-      case 'radio':
-        return Icons.radio;
-      case 'radio_button_checked':
-        return Icons.radio_button_checked;
-      case 'radio_button_unchecked':
-        return Icons.radio_button_unchecked;
-      case 'rate_review':
-        return Icons.rate_review;
-      case 'receipt':
-        return Icons.receipt;
-      case 'recent_actors':
-        return Icons.recent_actors;
-      case 'record_voice_over':
-        return Icons.record_voice_over;
-      case 'redeem':
-        return Icons.redeem;
-      case 'redo':
-        return Icons.redo;
-      case 'refresh':
-        return Icons.refresh;
-      case 'remove':
-        return Icons.remove;
-      case 'remove_circle':
-        return Icons.remove_circle;
-      case 'remove_circle_outline':
-        return Icons.remove_circle_outline;
-      case 'remove_from_queue':
-        return Icons.remove_from_queue;
-      case 'remove_red_eye':
-        return Icons.remove_red_eye;
-      case 'remove_shopping_cart':
-        return Icons.remove_shopping_cart;
-      case 'reorder':
-        return Icons.reorder;
-      case 'repeat':
-        return Icons.repeat;
-      case 'repeat_one':
-        return Icons.repeat_one;
-      case 'replay':
-        return Icons.replay;
-      case 'replay_5':
-        return Icons.replay_5;
-      case 'replay_10':
-        return Icons.replay_10;
-      case 'replay_30':
-        return Icons.replay_30;
-      case 'reply':
-        return Icons.reply;
-      case 'reply_all':
-        return Icons.reply_all;
-      case 'report':
-        return Icons.report;
-      case 'report_off':
-        return Icons.report_off;
-      case 'report_problem':
-        return Icons.report_problem;
-      case 'restaurant':
-        return Icons.restaurant;
-      case 'restaurant_menu':
-        return Icons.restaurant_menu;
-      case 'restore':
-        return Icons.restore;
-      case 'restore_from_trash':
-        return Icons.restore_from_trash;
-      case 'restore_page':
-        return Icons.restore_page;
-      case 'ring_volume':
-        return Icons.ring_volume;
-      case 'room':
-        return Icons.room;
-      case 'room_service':
-        return Icons.room_service;
-      case 'rotate_90_degrees_ccw':
-        return Icons.rotate_90_degrees_ccw;
-      case 'rotate_left':
-        return Icons.rotate_left;
-      case 'rotate_right':
-        return Icons.rotate_right;
-      case 'rounded_corner':
-        return Icons.rounded_corner;
-      case 'router':
-        return Icons.router;
-      case 'rowing':
-        return Icons.rowing;
-      case 'rss_feed':
-        return Icons.rss_feed;
-      case 'rv_hookup':
-        return Icons.rv_hookup;
-      case 'satellite':
-        return Icons.satellite;
-      case 'save':
-        return Icons.save;
-      case 'save_alt':
-        return Icons.save_alt;
-      case 'scanner':
-        return Icons.scanner;
-      case 'scatter_plot':
-        return Icons.scatter_plot;
-      case 'schedule':
-        return Icons.schedule;
-      case 'school':
-        return Icons.school;
-      case 'score':
-        return Icons.score;
-      case 'screen_lock_landscape':
-        return Icons.screen_lock_landscape;
-      case 'screen_lock_portrait':
-        return Icons.screen_lock_portrait;
-      case 'screen_lock_rotation':
-        return Icons.screen_lock_rotation;
-      case 'screen_rotation':
-        return Icons.screen_rotation;
-      case 'screen_share':
-        return Icons.screen_share;
-      case 'sd_card':
-        return Icons.sd_card;
-      case 'sd_storage':
-        return Icons.sd_storage;
-      case 'search':
-        return Icons.search;
-      case 'security':
-        return Icons.security;
-      case 'select_all':
-        return Icons.select_all;
-      case 'send':
-        return Icons.send;
-      case 'sentiment_dissatisfied':
-        return Icons.sentiment_dissatisfied;
-      case 'sentiment_neutral':
-        return Icons.sentiment_neutral;
-      case 'sentiment_satisfied':
-        return Icons.sentiment_satisfied;
-      case 'sentiment_very_dissatisfied':
-        return Icons.sentiment_very_dissatisfied;
-      case 'sentiment_very_satisfied':
-        return Icons.sentiment_very_satisfied;
-      case 'settings':
-        return Icons.settings;
-      case 'settings_applications':
-        return Icons.settings_applications;
-      case 'settings_backup_restore':
-        return Icons.settings_backup_restore;
-      case 'settings_bluetooth':
-        return Icons.settings_bluetooth;
-      case 'settings_brightness':
-        return Icons.settings_brightness;
-      case 'settings_cell':
-        return Icons.settings_cell;
-      case 'settings_ethernet':
-        return Icons.settings_ethernet;
-      case 'settings_input_antenna':
-        return Icons.settings_input_antenna;
-      case 'settings_input_component':
-        return Icons.settings_input_component;
-      case 'settings_input_composite':
-        return Icons.settings_input_composite;
-      case 'settings_input_hdmi':
-        return Icons.settings_input_hdmi;
-      case 'settings_input_svideo':
-        return Icons.settings_input_svideo;
-      case 'settings_overscan':
-        return Icons.settings_overscan;
-      case 'settings_phone':
-        return Icons.settings_phone;
-      case 'settings_power':
-        return Icons.settings_power;
-      case 'settings_remote':
-        return Icons.settings_remote;
-      case 'settings_system_daydream':
-        return Icons.settings_system_daydream;
-      case 'settings_voice':
-        return Icons.settings_voice;
-      case 'share':
-        return Icons.share;
-      case 'shop':
-        return Icons.shop;
-      case 'shop_two':
-        return Icons.shop_two;
-      case 'shopping_basket':
-        return Icons.shopping_basket;
-      case 'shopping_cart':
-        return Icons.shopping_cart;
-      case 'short_text':
-        return Icons.short_text;
-      case 'show_chart':
-        return Icons.show_chart;
-      case 'shuffle':
-        return Icons.shuffle;
-      case 'shutter_speed':
-        return Icons.shutter_speed;
-      case 'signal_cellular_4_bar':
-        return Icons.signal_cellular_4_bar;
-      case 'signal_cellular_connected_no_internet_4_bar':
-        return Icons.signal_cellular_connected_no_internet_4_bar;
-      case 'signal_cellular_no_sim':
-        return Icons.signal_cellular_no_sim;
-      case 'signal_cellular_null':
-        return Icons.signal_cellular_null;
-      case 'signal_cellular_off':
-        return Icons.signal_cellular_off;
-      case 'signal_wifi_4_bar':
-        return Icons.signal_wifi_4_bar;
-      case 'signal_wifi_4_bar_lock':
-        return Icons.signal_wifi_4_bar_lock;
-      case 'signal_wifi_off':
-        return Icons.signal_wifi_off;
-      case 'sim_card':
-        return Icons.sim_card;
-      case 'sim_card_alert':
-        return Icons.sim_card_alert;
-      case 'skip_next':
-        return Icons.skip_next;
-      case 'skip_previous':
-        return Icons.skip_previous;
-      case 'slideshow':
-        return Icons.slideshow;
-      case 'slow_motion_video':
-        return Icons.slow_motion_video;
-      case 'smartphone':
-        return Icons.smartphone;
-      case 'smoke_free':
-        return Icons.smoke_free;
-      case 'smoking_rooms':
-        return Icons.smoking_rooms;
-      case 'sms':
-        return Icons.sms;
-      case 'sms_failed':
-        return Icons.sms_failed;
-      case 'snooze':
-        return Icons.snooze;
-      case 'sort':
-        return Icons.sort;
-      case 'sort_by_alpha':
-        return Icons.sort_by_alpha;
-      case 'spa':
-        return Icons.spa;
-      case 'space_bar':
-        return Icons.space_bar;
-      case 'speaker':
-        return Icons.speaker;
-      case 'speaker_group':
-        return Icons.speaker_group;
-      case 'speaker_notes':
-        return Icons.speaker_notes;
-      case 'speaker_notes_off':
-        return Icons.speaker_notes_off;
-      case 'speaker_phone':
-        return Icons.speaker_phone;
-      case 'spellcheck':
-        return Icons.spellcheck;
-      case 'star':
-        return Icons.star;
-      case 'star_border':
-        return Icons.star_border;
-      case 'star_half':
-        return Icons.star_half;
-      case 'stars':
-        return Icons.stars;
-      case 'stay_current_landscape':
-        return Icons.stay_current_landscape;
-      case 'stay_current_portrait':
-        return Icons.stay_current_portrait;
-      case 'stay_primary_landscape':
-        return Icons.stay_primary_landscape;
-      case 'stay_primary_portrait':
-        return Icons.stay_primary_portrait;
-      case 'stop':
-        return Icons.stop;
-      case 'stop_screen_share':
-        return Icons.stop_screen_share;
-      case 'storage':
-        return Icons.storage;
-      case 'store':
-        return Icons.store;
-      case 'store_mall_directory':
-        return Icons.store_mall_directory;
-      case 'straighten':
-        return Icons.straighten;
-      case 'streetview':
-        return Icons.streetview;
-      case 'strikethrough_s':
-        return Icons.strikethrough_s;
-      case 'style':
-        return Icons.style;
-      case 'subdirectory_arrow_left':
-        return Icons.subdirectory_arrow_left;
-      case 'subdirectory_arrow_right':
-        return Icons.subdirectory_arrow_right;
-      case 'subject':
-        return Icons.subject;
-      case 'subscriptions':
-        return Icons.subscriptions;
-      case 'subtitles':
-        return Icons.subtitles;
-      case 'subway':
-        return Icons.subway;
-      case 'supervised_user_circle':
-        return Icons.supervised_user_circle;
-      case 'supervisor_account':
-        return Icons.supervisor_account;
-      case 'surround_sound':
-        return Icons.surround_sound;
-      case 'swap_calls':
-        return Icons.swap_calls;
-      case 'swap_horiz':
-        return Icons.swap_horiz;
-      case 'swap_horizontal_circle':
-        return Icons.swap_horizontal_circle;
-      case 'swap_vert':
-        return Icons.swap_vert;
-      case 'swap_vertical_circle':
-        return Icons.swap_vertical_circle;
-      case 'switch_camera':
-        return Icons.switch_camera;
-      case 'switch_video':
-        return Icons.switch_video;
-      case 'sync':
-        return Icons.sync;
-      case 'sync_disabled':
-        return Icons.sync_disabled;
-      case 'sync_problem':
-        return Icons.sync_problem;
-      case 'system_update':
-        return Icons.system_update;
-      case 'system_update_alt':
-        return Icons.system_update_alt;
-      case 'tab':
-        return Icons.tab;
-      case 'tab_unselected':
-        return Icons.tab_unselected;
-      case 'table_chart':
-        return Icons.table_chart;
-      case 'tablet':
-        return Icons.tablet;
-      case 'tablet_android':
-        return Icons.tablet_android;
-      case 'tablet_mac':
-        return Icons.tablet_mac;
-      case 'tag_faces':
-        return Icons.tag_faces;
-      case 'tap_and_play':
-        return Icons.tap_and_play;
-      case 'terrain':
-        return Icons.terrain;
-      case 'text_fields':
-        return Icons.text_fields;
-      case 'text_format':
-        return Icons.text_format;
-      case 'text_rotate_up':
-        return Icons.text_rotate_up;
-      case 'text_rotate_vertical':
-        return Icons.text_rotate_vertical;
-      case 'text_rotation_angledown':
-        return Icons.text_rotation_angledown;
-      case 'text_rotation_angleup':
-        return Icons.text_rotation_angleup;
-      case 'text_rotation_down':
-        return Icons.text_rotation_down;
-      case 'text_rotation_none':
-        return Icons.text_rotation_none;
-      case 'textsms':
-        return Icons.textsms;
-      case 'texture':
-        return Icons.texture;
-      case 'theaters':
-        return Icons.theaters;
-      case 'threed_rotation':
-        return Icons.threed_rotation;
-      case 'threesixty':
-        return Icons.threesixty;
-      case 'thumb_down':
-        return Icons.thumb_down;
-      case 'thumb_up':
-        return Icons.thumb_up;
-      case 'thumbs_up_down':
-        return Icons.thumbs_up_down;
-      case 'time_to_leave':
-        return Icons.time_to_leave;
-      case 'timelapse':
-        return Icons.timelapse;
-      case 'timeline':
-        return Icons.timeline;
-      case 'timer':
-        return Icons.timer;
-      case 'timer_3':
-        return Icons.timer_3;
-      case 'timer_10':
-        return Icons.timer_10;
-      case 'timer_off':
-        return Icons.timer_off;
-      case 'title':
-        return Icons.title;
-      case 'toc':
-        return Icons.toc;
-      case 'today':
-        return Icons.today;
-      case 'toll':
-        return Icons.toll;
-      case 'tonality':
-        return Icons.tonality;
-      case 'touch_app':
-        return Icons.touch_app;
-      case 'toys':
-        return Icons.toys;
-      case 'track_changes':
-        return Icons.track_changes;
-      case 'traffic':
-        return Icons.traffic;
-      case 'train':
-        return Icons.train;
-      case 'tram':
-        return Icons.tram;
-      case 'transfer_within_a_station':
-        return Icons.transfer_within_a_station;
-      case 'transform':
-        return Icons.transform;
-      case 'transit_enterexit':
-        return Icons.transit_enterexit;
-      case 'translate':
-        return Icons.translate;
-      case 'trending_down':
-        return Icons.trending_down;
-      case 'trending_flat':
-        return Icons.trending_flat;
-      case 'trending_up':
-        return Icons.trending_up;
-      case 'trip_origin':
-        return Icons.trip_origin;
-      case 'tune':
-        return Icons.tune;
-      case 'turned_in':
-        return Icons.turned_in;
-      case 'turned_in_not':
-        return Icons.turned_in_not;
-      case 'tv':
-        return Icons.tv;
-      case 'unarchive':
-        return Icons.unarchive;
-      case 'undo':
-        return Icons.undo;
-      case 'unfold_less':
-        return Icons.unfold_less;
-      case 'unfold_more':
-        return Icons.unfold_more;
-      case 'update':
-        return Icons.update;
-      case 'usb':
-        return Icons.usb;
-      case 'verified_user':
-        return Icons.verified_user;
-      case 'vertical_align_bottom':
-        return Icons.vertical_align_bottom;
-      case 'vertical_align_center':
-        return Icons.vertical_align_center;
-      case 'vertical_align_top':
-        return Icons.vertical_align_top;
-      case 'vibration':
-        return Icons.vibration;
-      case 'video_call':
-        return Icons.video_call;
-      case 'video_label':
-        return Icons.video_label;
-      case 'video_library':
-        return Icons.video_library;
-      case 'videocam':
-        return Icons.videocam;
-      case 'videocam_off':
-        return Icons.videocam_off;
-      case 'videogame_asset':
-        return Icons.videogame_asset;
-      case 'view_agenda':
-        return Icons.view_agenda;
-      case 'view_array':
-        return Icons.view_array;
-      case 'view_carousel':
-        return Icons.view_carousel;
-      case 'view_column':
-        return Icons.view_column;
-      case 'view_comfy':
-        return Icons.view_comfy;
-      case 'view_compact':
-        return Icons.view_compact;
-      case 'view_day':
-        return Icons.view_day;
-      case 'view_headline':
-        return Icons.view_headline;
-      case 'view_list':
-        return Icons.view_list;
-      case 'view_module':
-        return Icons.view_module;
-      case 'view_quilt':
-        return Icons.view_quilt;
-      case 'view_stream':
-        return Icons.view_stream;
-      case 'view_week':
-        return Icons.view_week;
-      case 'vignette':
-        return Icons.vignette;
-      case 'visibility':
-        return Icons.visibility;
-      case 'visibility_off':
-        return Icons.visibility_off;
-      case 'voice_chat':
-        return Icons.voice_chat;
-      case 'voicemail':
-        return Icons.voicemail;
-      case 'volume_down':
-        return Icons.volume_down;
-      case 'volume_mute':
-        return Icons.volume_mute;
-      case 'volume_off':
-        return Icons.volume_off;
-      case 'volume_up':
-        return Icons.volume_up;
-      case 'vpn_key':
-        return Icons.vpn_key;
-      case 'vpn_lock':
-        return Icons.vpn_lock;
-      case 'wallpaper':
-        return Icons.wallpaper;
-      case 'warning':
-        return Icons.warning;
-      case 'watch':
-        return Icons.watch;
-      case 'watch_later':
-        return Icons.watch_later;
-      case 'wb_auto':
-        return Icons.wb_auto;
-      case 'wb_cloudy':
-        return Icons.wb_cloudy;
-      case 'wb_incandescent':
-        return Icons.wb_incandescent;
-      case 'wb_iridescent':
-        return Icons.wb_iridescent;
-      case 'wb_sunny':
-        return Icons.wb_sunny;
-      case 'wc':
-        return Icons.wc;
-      case 'web':
-        return Icons.web;
-      case 'web_asset':
-        return Icons.web_asset;
-      case 'weekend':
-        return Icons.weekend;
-      case 'whatshot':
-        return Icons.whatshot;
-      case 'widgets':
-        return Icons.widgets;
-      case 'wifi':
-        return Icons.wifi;
-      case 'wifi_lock':
-        return Icons.wifi_lock;
-      case 'wifi_tethering':
-        return Icons.wifi_tethering;
-      case 'work':
-        return Icons.work;
-      case 'wrap_text':
-        return Icons.wrap_text;
-      case 'youtube_searched_for':
-        return Icons.youtube_searched_for;
-      case 'zoom_in':
-        return Icons.zoom_in;
-      case 'zoom_out':
-        return Icons.zoom_out;
-      case 'zoom_out_map':
-        return Icons.zoom_out_map;
-      default:
-        return null;
-    }
   }
 
   static bool truthful(value) {
